DFM Summary Methods

Summary and print methods for class 'dfm'. print.dfm just prints basic model information and the factor transition matrix [A], summary.dfm returns all system matrices and additional residual and goodness of fit statistics - with a print method allowing full or compact printout.

# S3 method for dfm
print(x, digits = 4L, ...)

# S3 method for dfm
summary(object, method = switch(object$em.method, none = "2s", "qml"), ...)

# S3 method for dfm_summary
print(x, digits = 4L, compact = sum(x$info["n"] > 15, x$info["n"] > 40), ...)

Arguments

x, object

an object class 'dfm'.

digits

integer. The number of digits to print out.

...

not used.

method

character. The factor estimates to use: one of "qml", "2s" or "pca".

compact

integer. Display a more compact printout: 0 prints everything, 1 omits the observation matrix [C] and covariance matrix [R], and 2 omits all disaggregated information - yielding a summary of only the factor estimates.

Value

Summary information following a dynamic factor model estimation.

Examples

mod = DFM(diff(BM14_Q), 2, 3)
#> Converged after 26 iterations.
print(mod)
#> Dynamic Factor Model: n = 9, T = 117, r = 2, p = 3, %NA = 7.5973
#> 
#> Factor Transition Matrix [A]
#>      L1.f1  L1.f2  L2.f1   L2.f2   L3.f1  L3.f2
#> f1  0.6426 0.5214 0.6945 -1.8621 -0.6978 1.1665
#> f2 -0.2663 1.0482 0.4616 -0.8371 -0.4214 0.4514
summary(mod)
#> Dynamic Factor Model: n = 9, T = 117, r = 2, p = 3, %NA = 7.5973
#> 
#> Call:  DFM(X = diff(BM14_Q), r = 2, p = 3)
#> 
#> Summary Statistics of Factors [F]
#>       N     Mean   Median      SD       Min     Max
#> f1  117  -0.0118   0.4768  2.8397  -18.1449  4.4612
#> f2  117  -0.0088  -0.0481  1.3171   -5.1039  2.5317
#> 
#> Factor Transition Matrix [A]
#>      L1.f1  L1.f2  L2.f1   L2.f2   L3.f1  L3.f2
#> f1  0.6426 0.5214 0.6945 -1.8621 -0.6978 1.1665
#> f2 -0.2663 1.0482 0.4616 -0.8371 -0.4214 0.4514
#> 
#> Factor Covariance Matrix [cov(F)]
#>          f1       f2
#> f1  8.0641   0.8894*
#> f2  0.8894*  1.7349 
#> 
#> Factor Transition Error Covariance Matrix [Q]
#>        u1     u2
#> u1 2.9293 0.9844
#> u2 0.9844 0.6969
#> 
#> Observation Matrix [C]
#>                 f1      f2
#> gdp         0.3192  0.0170
#> priv_cons   0.2211 -0.1466
#> invest      0.3019 -0.0710
#> export      0.2995  0.1821
#> import      0.3199  0.0628
#> empl        0.3069 -0.4699
#> prductivity 0.1683  0.1787
#> capacity    0.2330 -0.0046
#> gdp_us      0.1972  0.1042
#> 
#> Observation Error Covariance Matrix [diag(R) - Restricted]
#>         gdp   priv_cons      invest      export      import        empl 
#>      0.1583      0.6148      0.2842      0.1094      0.1221      0.0829 
#> prductivity    capacity      gdp_us 
#>      0.1380      0.3888      0.6212 
#> 
#> Observation Residual Covariance Matrix [cov(resid(DFM))]
#>                   gdp priv_cons    invest    export    import      empl
#> gdp           0.1345    0.1126*   0.0883*  -0.0497*  -0.0594*  -0.0064 
#> priv_cons     0.1126*   0.6103    0.0935*  -0.0814*  -0.0329   -0.0460*
#> invest        0.0883*   0.0935*   0.2678   -0.0759*  -0.0583*  -0.0150 
#> export       -0.0497*  -0.0814*  -0.0759*   0.0742    0.0083    0.0084 
#> import       -0.0594*  -0.0329   -0.0583*   0.0083    0.0953   -0.0061 
#> empl         -0.0064   -0.0460*  -0.0150    0.0084   -0.0061    0.0457 
#> prductivity   0.0898*   0.0925*   0.0790*  -0.0414*  -0.0824*  -0.0280*
#> capacity     -0.0526*  -0.0951   -0.0354   -0.0438*   0.0243   -0.0007 
#> gdp_us       -0.0354   -0.0044   -0.0065   -0.0209   -0.0418    0.0109 
#>             prductivity  capacity    gdp_us
#> gdp             0.0898*  -0.0526*  -0.0354 
#> priv_cons       0.0925*  -0.0951   -0.0044 
#> invest          0.0790*  -0.0354   -0.0065 
#> export         -0.0414*  -0.0438*  -0.0209 
#> import         -0.0824*   0.0243   -0.0418 
#> empl           -0.0280*  -0.0007    0.0109 
#> prductivity     0.1712   -0.0768*  -0.0540 
#> capacity       -0.0768*   0.4291   -0.0252 
#> gdp_us         -0.0540   -0.0252    0.6118 
#> 
#> Residual AR(1) Serial Correlation
#>         gdp   priv_cons      invest      export      import        empl 
#>    -0.14502    -0.16018    -0.13384    -0.01411    -0.04152    -0.23173 
#> prductivity    capacity      gdp_us 
#>    -0.09700     0.11880     0.15130 
#> 
#> Goodness of Fit: R-Squared
#>         gdp   priv_cons      invest      export      import        empl 
#>      0.8655      0.3897      0.7322      0.9258      0.9047      0.9543 
#> prductivity    capacity      gdp_us 
#>      0.8288      0.5709      0.3882 
#> 
#> Summary of Individual R-Squared's
#>   N    Mean  Median      SD     Min     Max
#>   9  0.7289  0.8288  0.2251  0.3882  0.9543