Dynamic Factor Models

dfms provides efficient estimation of Dynamic Factor Models via the EM Algorithm.

Details

Estimation can be done in 3 different ways following:

Doz, C., Giannone, D., & Reichlin, L. (2011). A two-step estimator for large approximate dynamic factor models based on Kalman filtering. Journal of Econometrics, 164(1), 188-205. doi:10.1016/j.jeconom.2011.02.012

Doz, C., Giannone, D., & Reichlin, L. (2012). A quasi-maximum likelihood approach for large, approximate dynamic factor models. Review of Economics and Statistics, 94(4), 1014-1024. doi:10.1162/REST_a_00225

Banbura, M., & Modugno, M. (2014). Maximum likelihood estimation of factor models on datasets with arbitrary pattern of missing data. Journal of Applied Econometrics, 29(1), 133-160. doi:10.1002/jae.2306

The default is em.method = "auto", which chooses "BM" following Banbura & Modugno (2014) with missing data or mixed frequency, and "DGR" following Doz, Giannone & Reichlin (2012) otherwise. Using em.method = "none" generates Two-Step estimates following Doz, Giannone & Reichlin (2011). This is extremely efficient on bigger datasets. PCA and Two-Step estimates are also reported in EM-estimation. All methods support missing data, but em.method = "DGR" does not model them in EM iterations.

Package Contents

Functions to Specify/Estimate Model and Key Methods

ICr() — Information Criteria

plot(<ICr>)

screeplot(<ICr>)

DFM() — Estimate the Model

summary(<dfm>)

plot(<dfm>)

as.data.frame(<dfm>)

residuals(<dfm>)

fitted(<dfm>)

predict(<dfm>) — Generate Forecasts

plot(<dfm_forecast>)

as.data.frame(<dfm_forecast>)

Auxiliary Functions

.VAR() — Estimate Vector Autoregression
SKF() — Stationary Kalman Filter
FIS() — Fixed Interval Smoother
SKFS() — Stationary Kalman Filter + Smoother
tsnarmimp() — Remove and Impute Missing Values in a Multivariate Time Series
ainv() — Rcpp Armadillo's Inverse Function
apinv() — Rcpp Armadillo's Pseudo-Inverse Function

Data

BM14_M — Monthly Series by Banbura and Modugno (2014)
BM14_Q — Quarterly Series by Banbura and Modugno (2014)
BM14_Models — Series Metadata + Small/Medium/Large Model Specifications