DFM Summary Methods

Summary and print methods for class 'dfm'. print.dfm just prints basic model information and the factor transition matrix [A], summary.dfm returns all system matrices and additional residual and goodness of fit statistics - with a print method allowing full or compact printout.

# S3 method for dfm
print(x, digits = 4L, ...)

# S3 method for dfm
summary(object, method = switch(object$em.method, none = "2s", "qml"), ...)

# S3 method for dfm_summary
print(x, digits = 4L, compact = sum(x$info["n"] > 15, x$info["n"] > 40), ...)

Arguments

x, object

an object class 'dfm'.

digits

integer. The number of digits to print out.

...

not used.

method

character. The factor estimates to use: one of "qml", "2s" or "pca".

compact

integer. Display a more compact printout: 0 prints everything, 1 omits the observation matrix [C] and covariance matrix [R], and 2 omits all disaggregated information - yielding a summary of only the factor estimates.

Value

Summary information following a dynamic factor model estimation.